requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
Django==4.1.2
djangorestframework==3.14.0
exceptiongroup==1.0.0rc9
h11==0.14.0
idna==3.4
lxml==4.9.1
outcome==1.2.0
pycparser==2.21
PySocks==1.7.1
pytz==2022.5
requests==2.28.1
selenium==4.5.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
trio==0.22.0
trio-websocket==0.9.2
TwoCaptcha==0.0.1
tzdata==2022.5
urllib3==1.26.12
wsproto==1.2.0
